requirements.txt
discord=2.3.2
Pillow=10.0.0
PySide6=6.5.3
gspread=5.11.2
clipboard=0.0.4
cryptography=41.0.4